Update generate_utils.py
I modified 'load_tracks' for taking into account the nontrivial indexing of xs, ys, zs. 
xs0, ys0, zs0 are the starting points both for signal and background, while previously we shifted by arr[:, 0] which is the starting point of the background events but not of signals, because we called 'cat_fn' for them.